+++ client/src/components/pages/ChatListPage/ChatListPage.js
@@ -1,0 +1,8 @@
+import React from "react"
+import ChatList from "./ChatList/ChatList"
+import withAuthRedirect from "hooks/useAuthRedirect/useAuthRedirect"
+
+const ChatListPage = () => <ChatList />
+
+const isAuthPage = false
+export default withAuthRedirect(ChatListPage, isAuthPage)
