+++ client/src/components/pages/HomePage/Swiper/SwipeButtons/SwipeButtons.test.js
