+++ client/src/components/pages/HomePage/Swiper/Swiper.test.js
