+++ client/src/components/pages/HomePage/Swiper/UserInfo/UserInfo.test.js
