+++ client/src/components/pages/WelcomePage/WelcomePage.js
