+++ client/src/components/pages/index.js
@@ -1,8 +1,8 @@
-export { default as WelcomePage } from "./WelcomePage"
-export { default as MissingPage } from "./MissingPage"
-export { default as LoginPage } from "./LoginPage"
-export { default as SignupPage } from "./SignupPage"
-export { default as HomePage } from "./HomePage"
-export { default as ChatPage } from "./ChatPage"
-export { default as ChatListPage } from "./ChatListPage"
-export { default as ProfilePage } from "./ProfilePage"
+export { default as WelcomePage } from "./WelcomePage/WelcomePage"
+export { default as MissingPage } from "./MissingPage/MissingPage"
+export { default as LoginPage } from "./LoginPage/LoginPage"
+export { default as SignupPage } from "./SignupPage/SignupPage"
+export { default as HomePage } from "./HomePage/HomePage"
+export { default as ChatPage } from "./ChatPage/ChatPage"
+export { default as ChatListPage } from "./ChatListPage/ChatListPage"
+export { default as ProfilePage } from "./ProfilePage/ProfilePage"
